requirements.txt
wordcloud>=1.8.1
Pillow>=9.0.1
nltk>= 3.6.1
plotly>=5.6.0
numpy>=1.21.2
pandas>=1.3.5
streamlit>=1.1.0